+++ spec/rubocop/cop/layout/empty_lines_after_module_inclusion_spec.rb
@@ -94,6 +94,26 @@ RSpec.describe RuboCop::Cop::Layout::EmptyLinesAfterModuleInclusion, :config do
       RUBY
     end
 
+    it "registers an offense and corrects when #{method} has multiple arguments" do
+      expect_offense(<<~RUBY, method: method)
+        class Foo
+          #{method} Bar, Baz
+          ^{method}^^^^^^^^^ Add an empty line after module inclusion.
+          def do_something
+          end
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        class Foo
+          #{method} Bar, Baz
+
+          def do_something
+          end
+        end
+      RUBY
+    end
+
     it "registers an offense and corrects for code that immediately follows #{method} inside a class" do
       expect_offense(<<~RUBY, method: method)
         class Bar
@@ -183,6 +203,30 @@ RSpec.describe RuboCop::Cop::Layout::EmptyLinesAfterModuleInclusion, :config do
       RUBY
     end
 
+    it "does not register an offense when #{method} is used with block method" do
+      expect_no_offenses(<<~RUBY)
+        #{method} Module.new do |arg|
+          do_something(arg)
+        end
+      RUBY
+    end
+
+    it "does not register an offense when #{method} is used with numbered block method", :ruby27 do
+      expect_no_offenses(<<~RUBY)
+        #{method} Module.new do
+          do_something(_1)
+        end
+      RUBY
+    end
+
+    it "does not register an offense when #{method} is used with `it` block method", :ruby34 do
+      expect_no_offenses(<<~RUBY)
+        #{method} Module.new do
+          do_something(it)
+        end
+      RUBY
+    end
+
     it "does not register an offense when using #{method} in `if` ... `else` branches" do
       expect_no_offenses(<<~RUBY)
         if condition
@@ -214,6 +258,26 @@ RSpec.describe RuboCop::Cop::Layout::EmptyLinesAfterModuleInclusion, :config do
     expect_no_offenses(<<~RUBY)
       it 'something' do
         expect(foo).to include(bar), "baz"
+      end
+    RUBY
+  end
+
+  it 'does not register an offense when `include` has zero arguments' do
+    expect_no_offenses(<<~RUBY)
+      class Foo
+        includes = [include, sdk_include].compact.map do |inc|
+          inc + "blah"
+        end.join(' ')
+      end
+    RUBY
+  end
+
+  it 'does not register an offense when module inclusion is called with modifier' do
+    expect_no_offenses(<<~RUBY)
+      class Foo
+        include Bar
+        include Baz if condition
+        include Qux
       end
     RUBY
   end