+++ spec/rubocop/cop/naming/method_name_spec.rb
@@ -99,6 +99,26 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
+    it 'accepts `alias_method` with non-intern first argument' do
+      expect_no_offenses(<<~RUBY)
+        alias_method foo, :bar
+        alias_method fooBar, :bar
+      RUBY
+    end
+
+    it 'accepts `alias_method` with array splat' do
+      expect_no_offenses(<<~RUBY)
+        alias_method *ary
+      RUBY
+    end
+
+    it 'accepts `alias_method` with unexpected arity' do
+      expect_no_offenses(<<~RUBY)
+        alias_method :foo, :bar, :baz
+        alias_method :fooBar, :bar, :baz
+      RUBY
+    end
+
     %w[class module].each do |kind|
       it "accepts class emitter method in a #{kind}" do
         expect_no_offenses(<<~RUBY)
@@ -315,6 +335,28 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
           expect_no_corrections
         end
       end
+
+      context 'for `alias` arguments' do
+        it 'registers an offense when member with forbidden name is defined' do
+          expect_offense(<<~RUBY, identifier: identifier)
+            alias %{identifier} foo
+                  ^{identifier} `%{identifier}` is forbidden, use another method name instead.
+          RUBY
+
+          expect_no_corrections
+        end
+      end
+
+      context 'for `alias_method` arguments' do
+        it 'registers an offense when member with forbidden name is defined' do
+          expect_offense(<<~RUBY, identifier: identifier)
+            alias_method :%{identifier}, :foo
+                         ^^{identifier} `%{identifier}` is forbidden, use another method name instead.
+          RUBY
+
+          expect_no_corrections
+        end
+      end
     end
   end
 
@@ -398,6 +440,28 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
           expect_no_corrections
         end
       end
+
+      context 'for `alias` arguments' do
+        it 'registers an offense when member with forbidden name is defined' do
+          expect_offense(<<~RUBY, identifier: identifier)
+            alias %{identifier} foo
+                  ^{identifier} `%{identifier}` is forbidden, use another method name instead.
+          RUBY
+
+          expect_no_corrections
+        end
+      end
+
+      context 'for `alias_method` arguments' do
+        it 'registers an offense when member with forbidden name is defined' do
+          expect_offense(<<~RUBY, identifier: identifier)
+            alias_method :%{identifier}, :foo
+                         ^^{identifier} `%{identifier}` is forbidden, use another method name instead.
+          RUBY
+
+          expect_no_corrections
+        end
+      end
     end
   end
 
@@ -429,6 +493,13 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       it 'does not register an offense when `define_method` is called with a variable`' do
         expect_no_offenses(<<~RUBY)
           #{name} foo do
+          end
+        RUBY
+      end
+
+      it 'does not register an offense when an operator method is defined using a string' do
+        expect_no_offenses(<<~RUBY)
+          #{name} '`' do
           end
         RUBY
       end
@@ -536,6 +607,27 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
+    it 'registers an offense for `alias` camelCase argument' do
+      expect_offense(<<~RUBY)
+        alias fooBar foo
+              ^^^^^^ Use snake_case for method names.
+      RUBY
+    end
+
+    it 'registers an offense for `alias_method` camelCase argument' do
+      expect_offense(<<~RUBY)
+        alias_method :fooBar, :foo
+                     ^^^^^^^ Use snake_case for method names.
+      RUBY
+    end
+
+    it 'registers an offense for `alias_method` snake_case string argument' do
+      expect_offense(<<~RUBY)
+        alias_method "fooBar", "foo"
+                     ^^^^^^^^ Use snake_case for method names.
+      RUBY
+    end
+
     include_examples 'never accepted',  'snake_case'
     include_examples 'always accepted', 'snake_case'
     include_examples 'multiple attr methods', 'snake_case'
@@ -618,7 +710,7 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
-    it 'registers an offense for `Struct` camelCase member' do
+    it 'registers an offense for `Struct` snake_case member' do
       expect_offense(<<~RUBY)
         Struct.new("foo_bar", var, *args, :snake_case, :camelCase, :snake_case_2, "camelCase2")
                                           ^^^^^^^^^^^ Use camelCase for method names.
@@ -626,7 +718,7 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
-    it 'registers an offense for `::Struct` camelCase member' do
+    it 'registers an offense for `::Struct` snake_case member' do
       expect_offense(<<~RUBY)
         ::Struct.new("foo_bar", var, *args, :snake_case, :camelCase, :snake_case_2, "camelCase2")
                                             ^^^^^^^^^^^ Use camelCase for method names.
@@ -634,7 +726,7 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
-    it 'registers an offense for `Data` camelCase member' do
+    it 'registers an offense for `Data` snake_case member' do
       expect_offense(<<~RUBY)
         Data.define(var, *args, :snake_case, :camelCase, :snake_case_2, "camelCase2")
                                 ^^^^^^^^^^^ Use camelCase for method names.
@@ -642,11 +734,32 @@ RSpec.describe RuboCop::Cop::Naming::MethodName, :config do
       RUBY
     end
 
-    it 'registers an offense for `::Data` camelCase member' do
+    it 'registers an offense for `::Data` snake_case member' do
       expect_offense(<<~RUBY)
         ::Data.define(var, *args, :snake_case, :camelCase, :snake_case_2, "camelCase2")
                                   ^^^^^^^^^^^ Use camelCase for method names.
                                                            ^^^^^^^^^^^^^ Use camelCase for method names.
+      RUBY
+    end
+
+    it 'registers an offense for `alias` snake_case argument' do
+      expect_offense(<<~RUBY)
+        alias foo_bar foo
+              ^^^^^^^ Use camelCase for method names.
+      RUBY
+    end
+
+    it 'registers an offense for `alias_method` snake_case symbol argument' do
+      expect_offense(<<~RUBY)
+        alias_method :foo_bar, :foo
+                     ^^^^^^^^ Use camelCase for method names.
+      RUBY
+    end
+
+    it 'registers an offense for `alias_method` snake_case string argument' do
+      expect_offense(<<~RUBY)
+        alias_method "foo_bar", "foo"
+                     ^^^^^^^^^ Use camelCase for method names.
       RUBY
     end